Move to engine and refactor
