+++ config/locales/en.yml
@@ -2841,6 +2841,8 @@ See the %{link} to find out more about %{sitename}'s features and to start using
 
   order_management:
     reports:
+      table:
+        select_and_search: "Select filters and click on %{option} to access your data."
       enterprise_fee_summaries:
         filters:
           date_range: "Date Range"
