+++ config/routes/api.rb
@@ -73,9 +73,5 @@ Openfoodnetwork::Application.routes.draw do
         end
       end
     end
-
-    namespace :reports do
-      get :packing
-    end
   end
 end
